requirements.txt
# SIU Object Detection Validator Requirements

# Deep Learning & Computer Vision
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Configuration & Utilities
pyyaml>=6.0.0
joblib>=1.3.0

# Logging & Progress
tqdm>=4.65.0
